Rename directory from shared-components to components
