+++ client/landing/stepper/declarative-flow/internals/steps-repository/components/setup-form/accent-color-control.tsx
