+++ client/landing/stepper/declarative-flow/internals/steps-repository/link-in-bio-post-setup/index.tsx
@@ -8,7 +8,7 @@ import FormattedHeader from 'calypso/components/formatted-header';
 import { SITE_STORE } from 'calypso/landing/stepper/stores';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import { useSite } from '../../../../hooks/use-site';
-import SetupForm from '../shared-components/setup-form';
+import SetupForm from '../components/setup-form';
 import type { Step } from '../../types';
 
 import '../link-in-bio-setup/styles.scss';
